Fix test for standalone membership
Made the wording of some membership steps membership-specific to execute
different code. That was needed because while the description of the
steps are the same, the underlying HTML differs. *Some* membership
feature steps are executed by donation ("frontend") step code. Maybe
this can be refactored in the future.

Deleted some unused elements from the membership page object because
they weren't used (because the donation steps are executed).